refactor(FeatureFlagDialog): consolidate, port to ts
