+++ client/src/dialogs/FeatureFlag/FeatureFlagDialog.scss
@@ -1,12 +1,19 @@
 .feature-flag-dialog {
-  width: 600px;
+  max-width: 600px;
+}
+
+/* Don't wrap the Creative Commons label. But this is a long link, so only
+  wrap on big screens. */
+@media only screen and (width >= 640px) {
+  .feature-flag-dialog ul {
+    column-count: 2;
+    column-gap: 2em;
+  }
 }
 
 .feature-flag-dialog ul {
   padding: 0;
   list-style-type: none;
-  column-count: 2;
-  column-gap: 2em;
 }
 
 .feature-flag-dialog ul li {
